+++ packages/mangrove.js/CHANGELOG.md
@@ -1,5 +1,7 @@
 # next version
 
+- Rename `penalty` to `bounty` in `Market.Summary` as it's a bounty from the taker's perspective
+
 # 0.11.1 (October 2022)
 
 - Fixed decimals handling for resting order in results.